tools/oblot-viewer/old/requirements.txt
future==0.18.2
lxml==4.5.2
numpy==1.19.1
Pillow==7.2.0
pymavlink==2.4.9
pyserial==3.4
six==1.15.0
wxPython==4.1.0
